+++ src/dom/DOMTokenList.h
@@ -44,7 +44,7 @@ public:
 
     static void tokenize(GCVector<String*>* tokens, String* src);
     static void concatTokensInsideParentheses(GCVector<String*>* tokens);
-    unsigned long length();
+    uint32_t length();
     Nullable<String*> item(unsigned long index);
     bool contains(String* token);
     String* addSingleToken(String* src, GCVector<String*>* tokens,
